requirements.txt
face_recognition==1.3.0
numpy==1.18.5
opencv_python==4.3.0.36
pandas==1.0.5
Pillow==9.3.0
streamlit==1.9.2
